TST: venv: Choose better package for --system-site-packages test
test_venv_system_site_packages() assumes that attrs isn't a package
that comes from the system.  That works for Travis with `dist:
xenial`, but it's not true for `dist: bionic` [0].  Plus, it may fail
locally depending on the user's system.

To reliably query a package that is not a system package, use the
custom python module that's created with
reproman.tests.utils.create_pymodule().

[0] https://travis-ci.org/github/ReproNim/reproman/builds/717727610#L1240