BF: DataladOrchestrator: Set DATALAD_SSH_IDENTITYFILE in __init__
DataladOrchestrator.prepare_remote() sets DATALAD_SSH_IDENTITYFILE so
that DataLad can work with SSH targets, in particular those for our
AwsEc2 resource, that need a custom key file that's not specified in
ssh_config.  Doing this in prepare_remote(), however, is problematic
because prepare_remote() isn't necessarily called before other methods
like fetch() that also require DATALAD_SSH_IDENTITYFILE to be
configured.  From the command line, this prevents fetching a job with
`reproman jobs` for a resource that depends on a custom key.

Move the DATALAD_SSH_IDENTITYFILE handling to __init__() so that it is
also in effect for resurrected orchestrators.

Fixes #561.