RF+BF: aws_ec2 - make more convenient about keys handling and more robust
- keys: big RF - no interactivity (unless fail)

First I just RFed functionality to provide a more helpful request for
a key_name.  On my laptop I ended up having over a dozen of
keys (since we do not remove them) and I decided to give specification
of an existing one as an option.  But then I just decided to change
behavior: is that if key_name is not specified, we do not even ask and
take resource/instance name as the name for the key which we might
either find locally or create a new one.

Another UX improvement is to not demand a key_filename.  If it is not
specified, match using expected location for that key_name.  I guess
it is useful to keep both key_name and key_filename available, so
people who already have some keys, could just specify them both.

- More robust State query: added doing multiple attempts to get State.
Encountered the case when initial call failed but subsequent
succeeded.

- AMI: switched to NITRC-CE.

The one which corresponds to v0.51.0 from Jan 09, 2020.  For some
reason newer v0.52.2 (from Dec, 2020) had older kernel...

find more here
https://aws.amazon.com/marketplace/server/configuration?productId=702680a0-e4b2-449a-9318-4ff1022d7381&ref_=psb_cfg_continue