RF+BF: move all keys relevant logic into AwsKeyMixin
Well, probably mixin is not really useful. Originally I thought to use it
also for AwsCondor but because it needs connection, ended up not doing that.
But nevertheless decided to keep as is for now at least.

Also added "a finite" loop to try create an instance while possibly asking for a new key