BF: orchestrators: Update for change in 'datalad subdatasets' key
In DataLad v0.12.0 (c89746c27e), the "revision" key in the results for
`datalad subdatasets` was changed from "revision" to "gitshasum", and
the old name was kept around for compatibility.  The old key will be
removed in the upcoming 0.14 release (f5c3104d88).

Look for "gitshasum", but fall back to "revision" because the DataLad
version on the remote may be below the version specified in setup.py
(currently 0.13).

Fixes #568.