TST: github ci: Restrict "on" to avoid duplicate runs
For this initial PR, the branch lives in the main repository rather
than a fork (which is more typical for this project) so that the
actions are triggered here and not in the fork.

This causes duplicate jobs, though, so for now restrict "push" to
pushes of the master branch.