TST: vcs: Choose a better "other branch" value
test_git_install_checkout() uses "other" as the name of a second
branch in addition to 'master'.  With Git's introduction of
init.defaultBranch, it's possible (though very unlikely) that the
default branch is set to "other", in which case this test will fail.

Avoid that by tacking a prefix onto the default branch name.