TST: github ci: Set git's init.defaultBranch
As of Git 2.28.0, the default branch for a freshly created repository
may be configured with init.defaultBranch.  Recent commits reworked a
few tests to not assume 'master' is the default branch.

Set up a custom init.defaultBranch value for the GitHub CI tests to
ensure that this assumption isn't reintroduced.  This also has the
benefit of silencing the warning 'git init' shows about
init.defaultBranch.