TST: orchestrators, vcs: Don't assume master is default branch
As of Git 2.28.0, the default branch for a freshly created repository
may be configured with init.defaultBranch.  Avoid hard coding 'master'
in test_orchestrators.py and test_vcs.py so that the tests will
continue to work if the user/system has init.defaultBranch set to a
value other than 'master' or if Git changes the default in a future
release.