MNT: Avoid invalid escape sequences in strings
A good number of regexps and a few other strings don't use an r-prefix
when the string contains non-escape-sequence backslashes.  This
doesn't make a functional difference at the moment because, when
Python encounters a backslash that produces an invalid escape
sequence, it automatically escapes the backslash.  But Python 3.6 and
later give a DeprecationWarning (<https://bugs.python.org/issue32912>),
with the goal of eventually making invalid escape sequences a syntax
error.