BF: debian: Drop spurious new line in version regexp
The re_versions regexp in parse_apt_cache_policy_pkgs_output() is
designed to match version output like this:

```
 *** 16.2.07~dfsg.1-2~nd90+1 500
        500 http://neuro.debian.net/debian stretch/contrib amd64 Packages
        100 /var/lib/dpkg/status
```

Following the part that matches priority on the first line ("500"),
the regexp includes "\n.*".  The regexp string is compiled with the
verbose flag and doesn't use an r prefix, so that "\n" is ignored,
leaving ".*" to slurp up any remaining characters on the first line.
It's not clear that the ".*" match is intentional or useful given the
preceding noop "\n" and that there are no test cases that require it,
so drop it along with the confusing "\n".

Note that if "\n" were kept around, the regexp would fail (according
to test_parse_apt_cache_policy_pkgs_output) when the next commit
converts this to a raw string because the unprocessed "\n" would no
longer be discarded by the verbose flag.