MNT: Use 'datalad push' instead of deprecated 'datalad publish'
'datalad publish' was marked as deprecated in 0.14 in favor of
'datalad push', which was added in 0.13 (ReproMan's current minimum
version).  For the purpose of prepare_remote(), the main differences
are 1) the '' value for --since must be spelled as '^', and 2) push()
must be called with data=auto to retain the behavior of the publish()
call (though going forward it's worth considering whether data=nothing
should be used instead).