RF: orchestrators: Avoid unnecessary publish() call
When preparing a DataLad remote, prepare_remote() calls publish() even
if the local HEAD is contained in the remote-tracking branch.  This
gives a notneeded result.  However, when the next commit switches this
over to push(), it will instead be an error with the form

   {'action': 'publish',
    'hints': None,
    'message': 'refs/heads/master->origin:refs/heads/master [rejected] (fetch '
               'first)',
    'operations': ['rejected', 'error'],
    'target': 'origin',
    ...}

Check whether the remote-tracking already has the commit and avoid
sending the changes if it does.  For publish(), this avoids an
unnecessary call.  For push(), this will side step the error in the
common case, though it's possible for the remote-tracking branch and
the branch in the remote to be out of sync in a way that makes the
push() fail, in particular if the remote branch has been reset to not
contain the local HEAD.  However, in that case, the error reported by
push() seems like a reasonable and clear one to propagate to the
caller.