BF: job_templates: Store launcher subjob output in expected files
Submit templates direct subjob output to
`$metadir/{stdout,stderr}.{0,..,$subjobs - 1}`.  The recently added
launcher handling, however, breaks this setup, creating just two files
with all the output (one for stdout and one for stderr).  In the case
of Slurm, the output files are `$metadir/{stdout,stderr}.4294967294`
because the files are constructed with "%a", which gets set to the
above value when there is no job array [1].

Rework launcher.template so that stdout and stderr are written to the
expected files.

[1] https://slurm.schedmd.com/job_array.html

Re: #578